components/producer/requirements.txt
Flask==2.3.3
requests==2.31.0
kafka-python==2.0.2
cassandra-driver==3.28.0
pytest==7.4.4
APScheduler==3.10.4
minio==7.1.17
